backend/requirements.txt
asgiref==3.11.1
boto3==1.42.46
botocore==1.42.46
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
dj-database-url==3.1.0
Django==5.2.11
django-cors-headers==4.9.0
django-ses==4.7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
jmespath==1.1.0
oauthlib==3.3.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.21.0
s3transfer==0.16.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
